% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-combine_nowcasts.R
\name{combine_nowcasts}
\alias{combine_nowcasts}
\title{Add a dadnow to a multidadnow object, or combine two multidadnow objects}
\usage{
combine_nowcasts(dadnow1, dadnow2)
}
\arguments{
\item{dadnow1, dadnow2}{A dadnow or multidadnow object.}
}
\value{
A single multidadnow object.
}
\description{
Add a dadnow to a multidadnow object, or combine two multidadnow objects
}
