% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-nowcast.R
\name{dispatch_model}
\alias{dispatch_model}
\title{Choose a model to use for nowcasting}
\usage{
dispatch_model(model)
}
\arguments{
\item{model}{The model to use for nowcasting. If not specified, the function attempts to find a function called \code{model}. See \code{vignette("Write_fit_XX_functions", package = "fatherStay")} for how to write your own model for use in this package.}
}
\description{
Choose a model to use for nowcasting
}
