% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-extract_results.R
\name{extract}
\alias{extract}
\title{Extract a model (or models) from a dadnow object}
\usage{
extract(dadnow, models)
}
\arguments{
\item{dadnow}{A dadnow object.}

\item{models}{A model name or vector of model names. Must correspond to the names of the models seen in the evaluations table. Note that the model names can change depending on the other models added to the dadnow object, so always double check.}
}
\value{
A dadnow object with the specified models.
}
\description{
Extract a model (or models) from a dadnow object
}
