% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-extract_results.R
\name{extract_best}
\alias{extract_best}
\title{Extract the best model (or models) from a dadnow object}
\usage{
extract_best(dadnow, metric = "rmse", top_n = 1)
}
\arguments{
\item{dadnow}{A dadnow object.}

\item{metric}{The metric to use for evaluation. Defaults to "rmse".}

\item{top_n}{The number of top models to return. Defaults to 1.}
}
\value{
A dadnow object with the specified models.
}
\description{
Extract the best model (or models) from a dadnow object
}
