% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_arima.R
\name{fit_arima}
\alias{fit_arima}
\title{Fit an ARIMA model on given data and make predictions for a given set of data with options for using an ARIMA or ARIMAX model}
\usage{
fit_arima(
  Y_train,
  X_train = NULL,
  X_nowcast = NULL,
  params = list(p = 1, d = 0, q = 1)
)
}
\arguments{
\item{params}{Named list containing additional parameters: \code{p} Integer indicating the number of parameters for the AR model, \code{d} Integer indicating the number of differences, \code{q} integer indicating the degree of the MA model}

\item{X_Train}{Training data for the explanatory variables in the model}

\item{Y_Train}{Training data for the response variable}

\item{X_Nowcast}{Current data of X_train for which there is no Y_train data, used to make nowcasts}
}
\value{
ARIMA model object, predictions, and fitted values of the model on the training data
}
\description{
Fit an ARIMA model on given data and make predictions for a given set of data with options for using an ARIMA or ARIMAX model
}
