% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_rf.R
\name{fit_rf}
\alias{fit_rf}
\title{Fit a Random Forest model on given data and make predictions for a given set of data}
\usage{
fit_rf(
  Y_train,
  X_train = NULL,
  X_nowcast = NULL,
  params = list(ntree = 500, mtry = NULL, weights = NULL, replace = TRUE, maxnodes =
    NULL, nodesize = 5)
)
}
\arguments{
\item{params}{Named list containing additional parameters:}

\item{X_Train}{Training data for the explanatory variables in the model}

\item{Y_Train}{Training data for the response variable}

\item{X_Nowcast}{Current data of X_train for which there is no Y_train data, used to make nowcasts}

\item{ntree}{Integer indicating the number of trees in the random forest, default is 500}

\item{mtry}{Indicates the number of features to try in each node of the Random Forest. Default is the default for randomForest}

\item{weights}{A vector of length same as y that are positive weights used only in sampling data to grow each tree (not used in any other calculation)}

\item{replace}{Should sampling of cases be done with or without replacement?}

\item{maxnodes}{Maximum number of terminal nodes trees in the forest can have.}
}
\value{
Random Forest object, predictions, and fitted values
}
\description{
Fit a Random Forest model on given data and make predictions for a given set of data
}
