% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_xgboost.R
\name{fit_xgboost}
\alias{fit_xgboost}
\title{Fit an eXtreme Gradient Boost model model on given data and make predictions for a given set of data}
\usage{
fit_xgboost(
  Y_train,
  X_train = NULL,
  X_nowcast = NULL,
  params = list(nrounds = 100, verbose = 1, XGBparams = list())
)
}
\arguments{
\item{params}{A named list containing options for the parameters used by \code{xgboost::xgb.train}: \code{nrounds} which is the number of boosting iteration to do, \code{XGBparams} a list of more parameters used in the model the most important of which is \code{max_depth} both \code{nrounds} and \code{max_depth} are tuned if their values are not specified, it also takes \code{verbose} should output be silent (0) or not (1)}

\item{X_Train}{Training data for the explanatory variables in the model}

\item{Y_Train}{Training data for the response variable}

\item{X_Nowcast}{Current data of X_train for which there is no Y_train data, used to make nowcasts}
}
\value{
XGBoost object, predictions, and fitted values
}
\description{
Fit an eXtreme Gradient Boost model model on given data and make predictions for a given set of data
}
