% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-get_attr.R
\name{get_data}
\alias{get_data}
\title{Get the data with predictions from a dadnow object}
\usage{
get_data(dadnow, include_training = TRUE)
}
\arguments{
\item{dadnow}{A dadnow object.}

\item{training}{A logical indicating whether to return the training data. If FALSE, then only the nowcasts are returned.}
}
\value{
A data frame with the predictions.
}
\description{
Get the data with predictions from a dadnow object
}
