% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-get_attr.R
\name{get_model}
\alias{get_model}
\title{Get the underlying model from a dadnow object}
\usage{
get_model(dadnow, model)
}
\arguments{
\item{dadnow}{A dadnow object.}
}
\value{
The underlying model, in it's original form.
}
\description{
Get the underlying model from a dadnow object
}
