% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-get_attr.R
\name{get_residuals}
\alias{get_residuals}
\title{Get the residuals from each model in a dadnow object}
\usage{
get_residuals(dadnow)
}
\arguments{
\item{dadnow}{A dadnow object.}
}
\value{
A data frame with the fitted values and residuals.
}
\description{
Get the residuals from each model in a dadnow object
}
