% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-prep_data.R
\name{prep_newdata}
\alias{prep_newdata}
\title{Prepare data for a particular formula, for use in predict(dadnow$models$model, newdata = X_train)}
\usage{
prep_newdata(formula, data, date_col = NULL, interpolate = TRUE)
}
\arguments{
\item{formula}{A formula object.}

\item{data}{A data frame.}

\item{date_col}{Name of the column containing date information. If NULL, the date information attempted to be inferred. If there's a single datetime column then it is used. If the data are a ts or mts or zoo object, the dates are esxtracted.}

\item{interpolate}{Whether to interpolate missing values. Defaults to TRUE.}
}
\value{
Data suitable for use in predict(dadnow$models$model, newdata = newdata)
}
\description{
This function prepares data in the same way as is done for the \code{nowcast()} function, but for use in \code{predict(dadnow$models$model, newdata = X_train)}. In particular, it ensures that the data is sorted by date, and that missing values are imputed, then parses the formula to create a model matrix (including lag terms).
}
